Create implementations for functions defined in header files
