+++ Code/GraphMol/RGroupDecomposition/CMakeLists.txt
@@ -1,5 +1,5 @@
 
-rdkit_library(RGroupDecomposition RGroupDecomp.cpp RGroupUtils.cpp RGroupCore.cpp
+rdkit_library(RGroupDecomposition RGroupDecomp.cpp RGroupDecompData.cpp RGroupData.cpp RGroupUtils.cpp RGroupCore.cpp
 				  RGroupDecompParams.cpp RGroupScore.cpp RGroupFingerprintScore.cpp RGroupGa.cpp
   LINK_LIBRARIES
   FMCS ChemTransforms SubstructMatch SmilesParse Fingerprints
@@ -28,7 +28,7 @@ rdkit_catch_test(rgroupCatchTests catch_rgd.cpp
 
 find_package(Boost ${RDK_BOOST_VERSION} COMPONENTS program_options)
 if(RDK_BUILD_CPP_TESTS AND Boost_FOUND)
-	add_executable(gaExample GaExample.cpp)
+	add_executable(gaExample GaExample.cpp RGroupDecompParams.h)
 	if(NOT Boost_USE_STATIC_LIBS)
 		target_compile_definitions(gaExample PUBLIC -DBOOST_PROGRAM_OPTIONS_DYN_LINK)
 	endif()
